requirements.txt
boto3
pylint
awscli
python-dotenv
dbt-snowflake